- Finish pvp implementation - Add migration to fix pvp default value - Redo spawnpoint saving logic and removed from Pokemon model - Fix spawnpoint merge - Fix typo parsing circle coordinates - Fix TLS protocol error thrown downloading game master file

+++ src/ChuckDeviceConfigurator/JobControllers/BaseSmartInstanceController.cs
@@ -102,7 +102,7 @@
                     // list, use 0 for route last completed sum calculation.
                     device.LastCompleted == 0 || device.LastCompletedWholeRoute == 0
                     ? 0
-                    : (double)device.LastCompleted - (double)device.LastCompletedWholeRoute
+                    : (double)device.LastCompleted - device.LastCompletedWholeRoute
                 );
                 if (totalRoundTime > 0)
                 {
